+++ Ecommerce.Backend/Ecommerce.sln
@@ -9,6 +9,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ecommerce.Repositorio", "Ec
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ecommerce.API", "Ecommerce.API\Ecommerce.API.csproj", "{EE8DBE92-CAAF-41B0-A662-D421BD0C8B98}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ecommerce.IoC", "Ecommerce.IoC\Ecommerce.IoC.csproj", "{BC6327A9-79FD-4B3A-822F-43A9D9C939D6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{EE8DBE92-CAAF-41B0-A662-D421BD0C8B98}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EE8DBE92-CAAF-41B0-A662-D421BD0C8B98}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EE8DBE92-CAAF-41B0-A662-D421BD0C8B98}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BC6327A9-79FD-4B3A-822F-43A9D9C939D6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BC6327A9-79FD-4B3A-822F-43A9D9C939D6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BC6327A9-79FD-4B3A-822F-43A9D9C939D6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BC6327A9-79FD-4B3A-822F-43A9D9C939D6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
